requirements.txt
click==6.7
pyusb==1.0.2
